Moving assignByString to lib.
