+++ packages/playground/src/constants/styleParams.js
@@ -4,6 +4,7 @@ import {
   NEW_PRESETS,
   defaultStyles,
   galleryOptions,
+  assignByString,
 } from 'pro-gallery-lib';
 
 const defaultStyleParams = defaultStyles;
@@ -94,22 +95,6 @@ export const getStyleParamsFromUrl = (locationSearchString) => {
     return {};
   }
 };
-function assignByString(
-  Obj,
-  string,
-  value,
-) {
-  let _obj = { ...Obj };
-  let keyArr = string.split('_');
-  let assignedProperty = keyArr.pop();
-  let pointer = _obj;
-  keyArr.forEach((key) => {
-    if (!pointer[key]) pointer[key] = {};
-    pointer = pointer[key];
-  });
-  pointer[assignedProperty] = value;
-  return _obj;
-}
 
 function flattenObject(ob) {
   var toReturn = {};
